Set up a "Sponsor us" page and set up jekyll layouts
Create a page that we can share with potential sponsors that holds the
most important info for them. In doing so I realized we were only using
plain HTML and CSS, which makes it tedious to add extra pages like this
one.

Github pages automatically compiles the site with Jekyll, meaning you
can use layouts and include partials. I pulled both the new and the old
design into a layout so you can use it on a site by setting the layout
in the YAML preamble at the top. Have a look at sponsor_us.md and
new_design.html to see how that works.