Change the phrasing of the target audience
Original discussion is here, https://github.com/clojurebridge-berlin/organization/issues/77

Also discussed at the core meeting.

This is a step in the right direction, but probably far from perfect. I've used a different approach for English and German, an explanatory footnote vs the FLTI acronym. The next step is getting feedback from people who are more acquainted with the subtleties of this issue.